requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
openai
httpx
python-docx==1.1.0
python-dotenv==1.0.0
pytesseract
Pillow
tiktoken
pymupdf
sqlalchemy[asyncio]
asyncpg 
python-multipart
aioboto3==13.1.0
streamlit
requests